home/requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.8.2
bitcoin-python==0.3
bootstrap4==0.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
config==0.5.0.post0
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
decouple==0.0.7
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.3
django-allauth==0.41.0
django-bootstrap4==1.1.1
django-channels==0.7.0
django-cleanup==5.0.0
django-cors-headers==3.3.0
django-countries==6.1.2
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-markdown-deux==1.0.5
django-pagedown==2.1.2
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.11.0
filelock==3.0.12
gamepy==0.1.5
heroku==0.1.4
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.5.0
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.1
Markdown==3.2.2
markdown2==2.3.9
mccabe==0.6.1
node==0.9.24
numpy==1.18.1
oauthlib==3.1.0
odict==1.7.0
parsel==1.6.0
paypal==1.2.5
Pillow==7.0.0
pipenv==2018.11.26
plumber==1.5
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pylint==2.4.4
pyOpenSSL==19.1.0
python-dateutil==1.5
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
queuelib==1.5.0
requests==2.23.0
requests-oauthlib==1.3.0
Scrapy==2.1.0
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
stripe==2.47.0
Twisted==20.3.0
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==20.0.5
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
w3lib==1.22.0
wrapt==1.11.2
zipp==3.0.0
zope.component==4.6
zope.deferredimport==4.3.1
zope.deprecation==4.4.0
zope.event==4.4
zope.hookable==5.0.0
zope.interface==4.7.1
zope.lifecycleevent==4.3
zope.proxy==4.3.3
